Split git_config into modules.
