+++ git-config/src/file/git_config/mod.rs
@@ -1,4 +1,8 @@
-pub mod includes;
+pub mod from_env;
+pub mod from_paths;
+mod resolve_includes;
+pub use from_env::functions::*;
+pub use resolve_includes::function::resolve_includes;
 
 use bstr::BStr;
 use std::{
@@ -6,10 +10,9 @@ use std::{
     collections::{HashMap, VecDeque},
     convert::TryFrom,
     fmt::Display,
-    path::{Path, PathBuf},
+    path::Path,
 };
 
-use crate::file::includes::resolve_includes;
 use crate::{
     file::{
         section::{MutableSection, SectionBody},
@@ -121,78 +124,6 @@ pub struct GitConfig<'event> {
     section_order: VecDeque<SectionId>,
 }
 
-pub mod from_paths {
-
-    use crate::{parser, values::path::interpolate};
-
-    /// The error returned by [`GitConfig::from_paths()`][super::GitConfig::from_paths()] and [`GitConfig::from_env_paths()`][super::GitConfig::from_env_paths()].
-    #[derive(Debug, thiserror::Error)]
-    #[allow(missing_docs)]
-    pub enum Error {
-        #[error(transparent)]
-        ParserOrIoError(#[from] parser::ParserOrIoError<'static>),
-        #[error(transparent)]
-        Interpolate(#[from] interpolate::Error),
-        #[error("The maximum allowed length {} of the file include chain built by following nested includes is exceeded", .max_depth)]
-        IncludeDepthExceeded { max_depth: u8 },
-        #[error("Include paths from environment variables must not be relative")]
-        MissingConfigPath,
-    }
-
-    /// Options when loading git config using [`GitConfig::from_paths()`][super::GitConfig::from_paths()].
-    #[derive(Clone, Copy)]
-    pub struct Options<'a> {
-        /// The location where gitoxide or git is installed
-        pub git_install_dir: Option<&'a std::path::Path>,
-        /// The maximum allowed length of the file include chain built by following nested includes where base level is depth = 0.
-        pub max_depth: u8,
-        /// When max depth is exceeded while following nested included, return an error if true or silently stop following
-        /// includes.
-        ///
-        /// Setting this value to false allows to read configuration with cycles, which otherwise always results in an error.
-        pub error_on_max_depth_exceeded: bool,
-        /// The location of the .git directory
-        pub git_dir: Option<&'a std::path::Path>,
-        /// The name of the branch that is currently checked out
-        pub branch_name: Option<&'a git_ref::FullNameRef>,
-    }
-
-    impl<'a> Default for Options<'a> {
-        fn default() -> Self {
-            Options {
-                git_install_dir: None,
-                max_depth: 10,
-                error_on_max_depth_exceeded: true,
-                git_dir: None,
-                branch_name: None,
-            }
-        }
-    }
-}
-
-pub mod from_env {
-    use super::from_paths;
-    use crate::values::path::interpolate;
-
-    /// Represents the errors that may occur when calling [`GitConfig::from_env`][crate::file::GitConfig::from_env()].
-    #[derive(Debug, thiserror::Error)]
-    #[allow(missing_docs)]
-    pub enum Error {
-        #[error("GIT_CONFIG_COUNT was not a positive integer: {}", .input)]
-        ParseError { input: String },
-        #[error("GIT_CONFIG_KEY_{} was not set", .key_id)]
-        InvalidKeyId { key_id: usize },
-        #[error("GIT_CONFIG_KEY_{} was set to an invalid value: {}", .key_id, .key_val)]
-        InvalidKeyValue { key_id: usize, key_val: String },
-        #[error("GIT_CONFIG_VALUE_{} was not set", .value_id)]
-        InvalidValueId { value_id: usize },
-        #[error(transparent)]
-        PathInterpolationError(#[from] interpolate::Error),
-        #[error(transparent)]
-        FromPathsError(#[from] from_paths::Error),
-    }
-}
-
 impl<'event> GitConfig<'event> {
     /// Constructs an empty `git-config` file.
     #[must_use]
@@ -247,112 +178,6 @@ impl<'event> GitConfig<'event> {
                 section_header.subsection_name,
                 other.sections.remove(&section_index).expect("present"),
             );
-        }
-    }
-
-    // TODO put into own module
-
-    /// Constructs a `git-config` from the default cascading sequence.
-    /// This is neither zero-alloc nor zero-copy.
-    ///
-    /// See <https://git-scm.com/docs/git-config#FILES> for details.
-    pub fn from_env_paths(options: from_paths::Options) -> Result<Self, from_paths::Error> {
-        use std::env;
-
-        let mut paths = vec![];
-
-        if env::var("GIT_CONFIG_NO_SYSTEM").is_err() {
-            if let Some(git_config_system) = env::var_os("GIT_CONFIG_SYSTEM") {
-                paths.push(PathBuf::from(git_config_system))
-            } else {
-                // In git the fallback is set to a build time macro which defaults to /etc/gitconfig
-                paths.push(PathBuf::from("/etc/gitconfig"));
-            }
-        }
-
-        if let Some(git_config_global) = env::var_os("GIT_CONFIG_GLOBAL") {
-            paths.push(PathBuf::from(git_config_global));
-        } else {
-            // Divergence from git-config(1)
-            // These two are supposed to share the same scope and override
-            // rather than append according to git-config(1) documentation.
-            if let Some(xdg_config_home) = env::var_os("XDG_CONFIG_HOME") {
-                paths.push(PathBuf::from(xdg_config_home).join("git/config"));
-            } else if let Some(home) = env::var_os("HOME") {
-                paths.push(PathBuf::from(home).join(".config/git/config"));
-            }
-
-            if let Some(home) = env::var_os("HOME") {
-                paths.push(PathBuf::from(home).join(".gitconfig"));
-            }
-        }
-
-        if let Some(git_dir) = env::var_os("GIT_DIR") {
-            paths.push(PathBuf::from(git_dir).join("config"));
-        }
-
-        Self::from_paths(paths, options)
-    }
-
-    /// Generates a config from the environment variables. This is neither
-    /// zero-copy nor zero-alloc. See [`git-config`'s documentation] on
-    /// environment variable for more information.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if `GIT_CONFIG_COUNT` set and is not a number, or if
-    /// there was an invalid key value pair.
-    ///
-    /// [`git-config`'s documentation]: https://git-scm.com/docs/git-config#Documentation/git-config.txt-GITCONFIGCOUNT
-    pub fn from_env(options: from_paths::Options) -> Result<Option<Self>, from_env::Error> {
-        use std::env;
-        let count: usize = match env::var("GIT_CONFIG_COUNT") {
-            Ok(v) => v.parse().map_err(|_| from_env::Error::ParseError { input: v })?,
-            Err(_) => return Ok(None),
-        };
-
-        let mut config = Self::new();
-        for i in 0..count {
-            let key =
-                env::var(format!("GIT_CONFIG_KEY_{}", i)).map_err(|_| from_env::Error::InvalidKeyId { key_id: i })?;
-            let value = env::var_os(format!("GIT_CONFIG_VALUE_{}", i))
-                .ok_or(from_env::Error::InvalidValueId { value_id: i })?;
-            if let Some((section_name, maybe_subsection)) = key.split_once('.') {
-                let (subsection, key) = if let Some((subsection, key)) = maybe_subsection.rsplit_once('.') {
-                    (Some(subsection), key)
-                } else {
-                    (None, maybe_subsection)
-                };
-
-                let mut section = if let Ok(section) = config.section_mut(section_name, subsection) {
-                    section
-                } else {
-                    // Need to have config own the section and subsection names
-                    // else they get dropped at the end of the loop.
-                    config.new_section(
-                        section_name.to_string(),
-                        subsection.map(|subsection| Cow::Owned(subsection.to_string())),
-                    )
-                };
-
-                section.push(
-                    Cow::<str>::Owned(key.to_string()).into(),
-                    Cow::Owned(git_path::into_bstr(PathBuf::from(value)).into_owned().into()),
-                );
-            } else {
-                return Err(from_env::Error::InvalidKeyValue {
-                    key_id: i,
-                    key_val: key.to_string(),
-                });
-            }
-        }
-
-        // This occurs when `GIT_CONFIG_COUNT` is set to zero.
-        if config.is_empty() {
-            Ok(None)
-        } else {
-            resolve_includes(&mut config, None, options)?;
-            Ok(Some(config))
         }
     }
